refactor: Tag 관련 Github 플랫폼 클래스 이름에 Tag 명시 (#217)
