+++ backend/pick-git/src/main/java/com/woowacourse/pickgit/tag/infrastructure/PlatformTagApiRequester.java
@@ -1,6 +1,6 @@
 package com.woowacourse.pickgit.tag.infrastructure;
 
-public interface PlatformApiRequester {
+public interface PlatformTagApiRequester {
 
     String requestTags(String url, String accessToken);
 }
