+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/unit/post/domain/PostTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import com.woowacourse.pickgit.exception.post.CannotAddTagException;
 import com.woowacourse.pickgit.post.domain.Post;
-import com.woowacourse.pickgit.post.domain.PostContent;
+import com.woowacourse.pickgit.post.domain.content.PostContent;
 import com.woowacourse.pickgit.tag.domain.Tag;
 import java.util.ArrayList;
 import java.util.Arrays;
