+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/unit/tag/TestTagConfiguration.java
@@ -1,7 +1,7 @@
 package com.woowacourse.pickgit.unit.tag;
 
 import com.woowacourse.pickgit.common.mockapi.MockTagApiRequester;
-import com.woowacourse.pickgit.tag.infrastructure.PlatformApiRequester;
+import com.woowacourse.pickgit.tag.infrastructure.PlatformTagApiRequester;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Bean;
 public class TestTagConfiguration {
 
     @Bean
-    public PlatformApiRequester platformApiRequester() {
+    public PlatformTagApiRequester platformTagApiRequester() {
         return new MockTagApiRequester();
     }
 }
